Start of avatar mocap example
