+++ engine/TestUtils.ts
@@ -1,7 +1,8 @@
 import { avatarPath } from '@etherealengine/common/src/schema.type.module'
-import { useHookstate } from '@etherealengine/hyperflux'
+import { NO_PROXY, useHookstate } from '@etherealengine/hyperflux'
 import { useFind } from '@etherealengine/spatial/src/common/functions/FeathersHooks'
 import { useEffect } from 'react'
+import { MathUtils } from 'three'
 
 export const useAvatars = () => {
   const avatars = useHookstate<string[]>([])
@@ -23,4 +24,17 @@ export const useAvatars = () => {
   }, [avatarQuery.status])
 
   return avatars
+}
+
+export const useRandomAvatar = () => {
+  const avatars = useAvatars()
+  const avatar = useHookstate<string | undefined>(undefined)
+
+  useEffect(() => {
+    const avatarArr = avatars.get(NO_PROXY)
+    if (!avatarArr.length || avatar.value) return
+    avatar.set(avatarArr[MathUtils.randInt(0, avatarArr.length)])
+  }, [avatars])
+
+  return avatar
 }
